chore(cron): move cron utilities to apps folder
