+++ apps/cron/src/auth/registry.ts
