+++ apps/cron/src/process-outbox/routes.ts
@@ -1,5 +1,5 @@
-import { cronAuth } from '@common/node/cron/auth/bearer';
 import express from 'express';
+import { cronAuth } from '../auth/bearer.ts';
 import * as controller from './controller.ts';
 
 export default express.Router().post('/process-outbox', cronAuth(), controller.process);
